+++ webapp/components/book/BookSlider.tsx
